chore(travis): skip build deploy if tag already exists on remote

+++ tools/scripts/release/publish-build-artifacts.sh
@@ -25,28 +25,34 @@ $(npm bin)/gulp build:release
 rm -rf $LOCAL_BUILDS_DIR
 mkdir -p $LOCAL_BUILDS_DIR
 
+BUILD_REPOS = "https://github.com/angular/flex-layout-builds"
+BUILD_EXISTS = $(git ls-remote --tags "$BUILD_REPOS" "$buildVersion-$commitSha")
 
-# Clone the repository
-git clone https://github.com/angular/flex-layout-builds \
-  $LOCAL_BUILDS_DIR --depth=2
+if [ "$BUILD_EXISTS" == "" ]; then
 
-# Copy the build files to the repository
-rm -rf $LOCAL_BUILDS_DIR/*
-cp -r $buildDir/* $LOCAL_BUILDS_DIR
-cp CHANGELOG.md $LOCAL_BUILDS_DIR
+  # Clone the repository
+  git clone $BUILD_REPOS \
+    $LOCAL_BUILDS_DIR --depth=2
 
-# Create the build commit and push the changes to the repository.
-cd $LOCAL_BUILDS_DIR
+  # Copy the build files to the repository
+  rm -rf $LOCAL_BUILDS_DIR/*
+  cp -r $buildDir/* $LOCAL_BUILDS_DIR
+  cp CHANGELOG.md $LOCAL_BUILDS_DIR
 
-# Prepare Git for pushing the artifacts to the repository.
-git config user.name "${commitAuthorName}"
-git config user.email "${commitAuthorEmail}"
-git config credential.helper "store --file=.git/credentials"
-echo "https://${FLEX_LAYOUT_BUILDS_TOKEN}:@github.com" > .git/credentials
+  # Create the build commit and push the changes to the repository.
+  cd $LOCAL_BUILDS_DIR
 
-git add -A
-git commit --allow-empty -m "build: $buildVersion-$commitSha"
-git tag "$buildVersion-$commitSha"
-git push -q origin master --tags
+  # Prepare Git for pushing the artifacts to the repository.
+  git config user.name "${commitAuthorName}"
+  git config user.email "${commitAuthorEmail}"
+  git config credential.helper "store --file=.git/credentials"
+  echo "https://$FLEX_LAYOUT_BUILDS_TOKEN:@github.com" > .git/credentials
 
-echo "Version '$buildVersion-$commitSha' pushed successfully to angular/flex-layout-builds!"
+  git add -A
+  git commit --allow-empty -m "build: $buildVersion-$commitSha"
+  git tag "$buildVersion-$commitSha"
+  git push -q origin master --tags
+  echo "Version '$buildVersion-$commitSha' pushed successfully to $BUILD_REPOS!"
+
+fi
+
